Aggregated changes
- Pedantic compiler flags in gcc and clang
- Explicit type conversions, define default constructors, auto where suggested, unused variables, no comma after last entry in enums, init for local variables, merge if-else blocks with same content
- Typos and grammer

Content:
- Brush.setColor()
- SurfaceLost exception
- Messy attempt for Android vulkan
- Vulkan validation layer changes also quite messy
- Vulkan pipeline layout destroyed after pipelines
- deviceList()
- SoundDevice
--- speaker config
--- enumerate devices
- Sound effect
--- OpenAL effect extension with some custom changes so each effect can have its individual reverb strength
--- EAX reverb
--- setDirection(), setVelocity(), setPitch(), setInnerConeAngle(), setOuterConeAngle(), setOuterConeGain
- RFile modification time
- Matrix4x4 - identiyMatrix(), projectChecked(), position(), yawPitchRoll() functions added
- SystemApi createWindow with application title
- Additional keys mapped - numpad and pageUp and a few others on windows and x11
- X11Api - set title for dock icon, no key auto-repeat routed to event-dispatcher, focus in/out events
- Timer stop() resets interval
- Widget absoluteRect()
- Several classes have key event overrides for Gamepad, pointer and appState, but not all
- Log for vector and size types
- Vector classes array access operator

Updated nanosvg, stb_image_write, stb_image, stb_truetype

+++ Engine/2d/brush.cpp
@@ -10,8 +10,8 @@ Brush::Brush(const Texture2d &texture, PaintDevice::Blend b, ClampMode c)
   :tex(texture),texFrm(texture.format()),color(1.f),blend(b),clamp(c) {
   info.w    = texture.w();
   info.h    = texture.h();
-  info.invW = 1.f/info.w;
-  info.invH = 1.f/info.h;
+  info.invW = 1.f/(float)info.w;
+  info.invH = 1.f/(float)info.h;
   }
 
 Brush::Brush(const Color &color, PaintDevice::Blend b)
@@ -26,8 +26,8 @@ Brush::Brush(const Texture2d &texture, const Color &color, PaintDevice::Blend b,
   :tex(texture),texFrm(texture.format()),color(color),blend(b),clamp(c) {
   info.w    = texture.w();
   info.h    = texture.h();
-  info.invW = 1.f/info.w;
-  info.invH = 1.f/info.h;
+  info.invW = 1.f/(float)info.w;
+  info.invH = 1.f/(float)info.h;
   }
 
 Brush::Brush(const Sprite &texture, PaintDevice::Blend b)
@@ -36,10 +36,10 @@ Brush::Brush(const Sprite &texture, PaintDevice::Blend b)
 
   info.w    = texture.w();
   info.h    = texture.h();
-  info.invW = 1.f/r.w;
-  info.invH = 1.f/r.h;
-  info.dx   = r.x*info.invW;
-  info.dy   = r.y*info.invH;
+  info.invW = 1.f/(float)r.w;
+  info.invH = 1.f/(float)r.h;
+  info.dx   = (float)r.x*info.invW;
+  info.dy   = (float)r.y*info.invH;
   }
 
 Brush::Brush(const Sprite &texture, const Color &color, PaintDevice::Blend b)
@@ -48,8 +48,8 @@ Brush::Brush(const Sprite &texture, const Color &color, PaintDevice::Blend b)
 
   info.w    = texture.w();
   info.h    = texture.h();
-  info.invW = 1.f/r.w;
-  info.invH = 1.f/r.h;
-  info.dx   = r.x*info.invW;
-  info.dy   = r.y*info.invH;
+  info.invW = 1.f/(float)r.w;
+  info.invH = 1.f/(float)r.h;
+  info.dx   = (float)r.x*info.invW;
+  info.dy   = (float)r.y*info.invH;
   }
